services planned

+++ src/main/java/Base/Repository/Repository.java
@@ -1,17 +1,17 @@
 package Base.Repository;
 
 import Base.Entity.BaseEntity;
+import utils.Exceptions.EntityMustUseUpdateInsteadOfSave;
+import utils.Exceptions.EntityNotFoundException;
 import utils.JPA;
 
+import javax.persistence.EntityExistsException;
 import javax.persistence.EntityManager;
-import javax.persistence.EntityManagerFactory;
 import javax.persistence.EntityTransaction;
-import javax.persistence.Persistence;
 import java.io.Serializable;
 import java.util.Collection;
-import java.util.Optional;
 
-public abstract class Repository<E extends BaseEntity<ID>,ID extends Serializable> implements RepositoryApi<E,ID> {
+public abstract class Repository<E extends BaseEntity<ID>, ID extends Serializable> implements RepositoryApi<E, ID> {
 
     private final EntityManager entityManager;
 
@@ -22,23 +22,21 @@ public abstract class Repository<E extends BaseEntity<ID>,ID extends Serializabl
     }
 
 
-
-
     public EntityTransaction getTransaction() {
-        if(entityManager.isOpen())
+        if (entityManager.isOpen())
             return entityManager.getTransaction();
         else throw new IllegalStateException();
     }
 
     public EntityManager getEntityManager() {
-        if(entityManager.isOpen())
+        if (entityManager.isOpen())
             return entityManager;
         else throw new IllegalStateException();
     }
 
     @Override
     public void save(E e) throws Exception {
-        if(!entityManager.isOpen())
+        if (!entityManager.isOpen())
             throw new IllegalStateException();
         if (e.getId() == null) {
             EntityTransaction transaction = entityManager.getTransaction();
@@ -51,7 +49,28 @@ public abstract class Repository<E extends BaseEntity<ID>,ID extends Serializabl
                 exception.printStackTrace();
 
             }
+        } else throw new
+                EntityMustUseUpdateInsteadOfSave(
+                "the entities with id must be updated not saved"
+        );
+//        EntityTransaction transaction = entityManager.getTransaction();
+//        try {
+//            transaction.begin();
+//            entityManager.merge(e);
+//            transaction.commit();
+//        } catch (Exception exception) {
+//            if (transaction.isActive()) transaction.rollback();
+//        }
+    }
+
+
+    @Override
+    public void update(E e) throws Exception {
+        E elementInDb = entityManager.find(getType(), e.getId());
+        if (e.getId() == null || elementInDb == null) {
+            throw new EntityNotFoundException(" the entity does not exist in database ");
         }
+
         EntityTransaction transaction = entityManager.getTransaction();
         try {
             transaction.begin();
@@ -63,15 +82,21 @@ public abstract class Repository<E extends BaseEntity<ID>,ID extends Serializabl
     }
 
     @Override
-    public void delete(E e) {
+    public void delete(E e) throws EntityNotFoundException {
+        E elementInDB = entityManager.find(getType(), e.getId());
+        if (elementInDB == null) {
+            throw new EntityNotFoundException(" the entity does not exist in database ");
+        }
         entityManager.getTransaction().begin();
         entityManager.remove(e);
         entityManager.getTransaction().commit();
     }
 
     @Override
-    public Optional<E> getById(ID id) {
-        return Optional.ofNullable(entityManager.find(getType(), id));
+    public E getById(ID id) throws Exception {
+        E entity = entityManager.find(getType(), id);
+        if (entity == null) throw new EntityNotFoundException(" the entity does not exist in database ");
+        return entity;
     }
 
     @Override
@@ -88,5 +113,8 @@ public abstract class Repository<E extends BaseEntity<ID>,ID extends Serializabl
 
     }
 
-
+    @Override
+    public String name() {
+        return getType().getSimpleName();
+    }
 }
